split unified job templates out into jts and projects because workflows don't have custom virtualenvs & rename file so the command is shorter
update copypasta and add sanity check
